+++ tests/assert/database/migrations/20181031094608_composite_id.php
@@ -55,8 +55,8 @@ class CompositeId extends AbstractMigration
     {
         $sql = <<<'EOT'
             CREATE TABLE `composite_id` (
-                `id1` int(11) NOT NULL DEFAULT 0 COMMENT 'ID 1',
-                `id2` int(11) NOT NULL DEFAULT 0 COMMENT 'ID 2',
+                `id1` bigint(20) NOT NULL DEFAULT 0 COMMENT 'ID 1',
+                `id2` bigint(20) NOT NULL DEFAULT 0 COMMENT 'ID 2',
                 `name` varchar(32) NOT NULL DEFAULT '' COMMENT '名字',
                 PRIMARY KEY (`id1`,`id2`)
                 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COMMENT='带复合主键的表';
