+++ tests/assert/database/migrations/20190424055915_test_unique.php
@@ -55,7 +55,7 @@ class TestUnique extends AbstractMigration
     {
         $sql = <<<'EOT'
             CREATE TABLE `test_unique` (
-                `id` int(11) NOT NULL AUTO_INCREMENT COMMENT 'ID',
+                `id` bigint(20) NOT NULL AUTO_INCREMENT COMMENT 'ID',
                 `name` varchar(64) NOT NULL DEFAULT '' COMMENT '名字',
                 `create_at` datetime NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT '创建时间',
                 `identity` varchar(64) NOT NULL DEFAULT '' COMMENT '唯一标识符',
