+++ tests/assert/database/migrations/20191005020220_user_role_soft_deleted.php
@@ -55,9 +55,9 @@ class UserRoleSoftDeleted extends AbstractMigration
     {
         $sql = <<<'EOT'
             CREATE TABLE `user_role_soft_deleted` (
-                `id` int(11) NOT NULL AUTO_INCREMENT COMMENT 'ID',
-                `user_id` int(11) NOT NULL DEFAULT '0' COMMENT '用户 ID',
-                `role_id` int(11) NOT NULL DEFAULT '0' COMMENT '角色 ID',
+                `id` bigint(20) NOT NULL AUTO_INCREMENT COMMENT 'ID',
+                `user_id` bigint(20) NOT NULL DEFAULT '0' COMMENT '用户 ID',
+                `role_id` bigint(20) NOT NULL DEFAULT '0' COMMENT '角色 ID',
                 `create_at` datetime NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT '创建时间',
                 `delete_at` bigint(20) unsigned NOT NULL DEFAULT '0' COMMENT '删除时间 0=未删除;大于0=删除时间;',
                 PRIMARY KEY (`id`)
